move around feed logic

+++ peanut/strand/rest_api_views.py
@@ -172,12 +172,8 @@ class RetrieveUpdateDestroyStrandInviteAPI(RetrieveUpdateDestroyAPIView):
             # Run through old actions to see if we need to change the timing of the join (incase the "add"
             #    action happened first).  Also remove if any old ones exist
             for oldAction in oldActions:
-                if oldAction.added < action.added:
-                    action.added = oldAction.added - datetime.timedelta(seconds=1)
-                    action.save()
-
                 # Can't join a strand more than once, just do a quick check for that
-                if oldAction.action_type == action.action_type:
+                if oldAction.action_type == action.action_type and oldAction.user == action.user:
                     action.delete()
 
 """
